Fixed value interpreted as a format string when using `t:key`
It previously assumed that using `t:key` without any variable argument would just use
the `translate:key` method but that was wrong.
So every call just using `t:key` ended up as being interpreted as a format string. Which is is wrong.
Since we cannot introspect the variable arguments we need an explicit method to use the short format syntax. 
That new method is `tf:key, ...`